backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
httpx>=0.27.0
networkx>=3.3
pandas>=2.2.2
numpy>=1.26.4
supabase>=2.4.5
python-dateutil>=2.9.0
pydantic>=2.7.0
scipy>=1.13.0
